Move index functions into include file.
